Implement service auth

+++ tests-api/src/main/java/com/voteva/tests/grpc/service/v1/impl/TestsServiceV1GrpcImpl.java
@@ -1,7 +1,9 @@
 package com.voteva.tests.grpc.service.v1.impl;
 
+import com.voteva.auth.grpc.model.v1.GCheckAccessRequest;
 import com.voteva.common.grpc.model.Empty;
 import com.voteva.tests.converter.ModelConverter;
+import com.voteva.tests.grpc.client.GRpcAccessServiceClient;
 import com.voteva.tests.grpc.model.v1.GAddTestRequest;
 import com.voteva.tests.grpc.model.v1.GAddTestResponse;
 import com.voteva.tests.grpc.model.v1.GGetAllTestsRequest;
@@ -15,6 +17,7 @@ import com.voteva.tests.grpc.model.v1.GGetTestsByCategoryResponse;
 import com.voteva.tests.grpc.model.v1.GRemoveTestRequest;
 import com.voteva.tests.grpc.service.v1.TestsServiceV1Grpc;
 import com.voteva.tests.model.entity.TestEntity;
+import com.voteva.tests.security.Grants;
 import com.voteva.tests.service.TestsService;
 import io.grpc.stub.StreamObserver;
 import org.lognet.springboot.grpc.GRpcService;
@@ -28,10 +31,15 @@ import java.util.stream.Collectors;
 @GRpcService
 public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1ImplBase {
 
+    private final GRpcAccessServiceClient accessServiceClient;
     private final TestsService testsService;
 
     @Autowired
-    public TestsServiceV1GrpcImpl(TestsService testsService) {
+    public TestsServiceV1GrpcImpl(
+            GRpcAccessServiceClient accessServiceClient,
+            TestsService testsService) {
+
+        this.accessServiceClient = accessServiceClient;
         this.testsService = testsService;
     }
 
@@ -39,6 +47,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
     public void getTestCategories(
             GGetTestCategoriesRequest request,
             StreamObserver<GGetTestCategoriesResponse> responseObserver) {
+
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_READ_GRANT.getValue())
+                .build());
 
         List<String> categories = testsService.getCategories();
 
@@ -52,6 +65,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
     public void getAllTests(
             GGetAllTestsRequest request,
             StreamObserver<GGetAllTestsResponse> responseObserver) {
+
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_READ_GRANT.getValue())
+                .build());
 
         Page<TestEntity> tests = testsService.getAllTests(
                 ModelConverter.convert(request.getPageable()));
@@ -71,6 +89,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
             GGetTestsByCategoryRequest request,
             StreamObserver<GGetTestsByCategoryResponse> responseObserver) {
 
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_READ_GRANT.getValue())
+                .build());
+
         Page<TestEntity> tests = testsService.getTestsByCategory(
                 request.getCategory(), ModelConverter.convert(request.getPageable()));
 
@@ -89,6 +112,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
             GGetTestRequest request,
             StreamObserver<GGetTestResponse> responseObserver) {
 
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_READ_GRANT.getValue())
+                .build());
+
         TestEntity testEntity = testsService.getTest(
                 ModelConverter.convert(request.getTestUid()));
 
@@ -103,6 +131,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
             GAddTestRequest request,
             StreamObserver<GAddTestResponse> responseObserver) {
 
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_WRITE_GRANT.getValue())
+                .build());
+
         UUID testUid = testsService.addTest(ModelConverter.convert(request));
 
         responseObserver.onNext(GAddTestResponse.newBuilder()
@@ -115,6 +148,11 @@ public class TestsServiceV1GrpcImpl extends TestsServiceV1Grpc.TestsServiceV1Imp
     public void removeTest(
             GRemoveTestRequest request,
             StreamObserver<Empty> responseObserver) {
+
+        accessServiceClient.checkAccess(GCheckAccessRequest.newBuilder()
+                .setAuthentication(request.getAuthentication())
+                .setGrant(Grants.TESTS_WRITE_GRANT.getValue())
+                .build());
 
         testsService.removeTest(ModelConverter.convert(request.getTestUid()));
